requirements.txt
torch==2.6.0+cu124
torchvision==0.17.0+cu124
torchaudio==2.6.0+cu124
--extra-index-url https://download.pytorch.org/whl/cu124
fvcore
xformers
torchmetrics
blenderproc
opencv-python 
omegaconf
ruamel.yaml
hydra-colorlog
hydra-core
triton
gdown
pandas
imageio
pyrender
pytorch-lightning
pycocotools 
distinctipy
git+https://github.com/facebookresearch/segment-anything.git
ultralytics
timm
gorilla-core
trimesh
gpustat
imgaug
einops
